@@ -4,7 +4,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build --prod --configuration=$CONFIG",
     "test": "ng test",
     "tests": "run-p test:*",
     "test:dapp": "ng test dapp --browsers=ChromeHeadless --watch=false --progress=false",
@@ -16,54 +16,70 @@
     "lint:all": "node ./lint-all.js",
     "commit": "lint-staged --allow-empty",
     "push": "run-s lint build e2e tests",
-    "heroku": "pm2-runtime start heroku.config.js"
+    "heroku": "pm2-runtime start heroku.config.js",
+    "postinstall": "run-p envsub:*",
+    "envsub:stage": "envsub --diff ./projects/dapp/src/environments/environment.stage.ts",
+    "envsub:prod": "envsub --diff ./projects/dapp/src/environments/environment.prod.ts"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.2.2",
-    "@angular/cdk": "^10.2.7",
-    "@angular/common": "~10.2.2",
-    "@angular/compiler": "~10.2.2",
-    "@angular/core": "~10.2.2",
-    "@angular/forms": "~10.2.2",
-    "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "~10.2.2",
-    "@angular/platform-browser-dynamic": "~10.2.2",
-    "@angular/router": "~10.2.2",
+    "@angular/animations": "~11.0.5",
+    "@angular/cdk": "^11.0.3",
+    "@angular/common": "~11.0.5",
+    "@angular/compiler": "~11.0.5",
+    "@angular/core": "~11.0.5",
+    "@angular/forms": "~11.0.5",
+    "@angular/material": "^11.0.3",
+    "@angular/platform-browser": "~11.0.5",
+    "@angular/platform-browser-dynamic": "~11.0.5",
+    "@angular/router": "~11.0.5",
     "@ngneat/transloco": "^2.19.1",
     "@ngxd/core": "^10.0.0",
+    "@types/lodash": "^4.14.165",
     "@waves.exchange/provider-web": "^0.2.3",
     "@waves/signer": "^0.1.2",
+    "envsub": "^4.0.7",
     "include-media": "^1.4.9",
+    "ngx-markdown": "^10.1.1",
+    "ngx-mask": "^11.0.1",
     "ngx-progressbar": "^6.0.3",
+    "npm-run-all": "^4.1.5",
     "pm2": "^4.5.0",
     "rxjs": "~6.6.3",
     "tslib": "^2.0.0",
+    "tslint-etc": "^1.13.9",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1002.0",
-    "@angular-devkit/build-ng-packagr": "~0.1002.0",
-    "@angular/cli": "~10.2.0",
-    "@angular/compiler-cli": "~10.2.2",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "~0.1100.5",
+    "@angular-eslint/builder": "1.0.0",
+    "@angular-eslint/eslint-plugin": "1.0.0",
+    "@angular-eslint/eslint-plugin-template": "1.0.0",
+    "@angular-eslint/schematics": "1.0.0",
+    "@angular-eslint/template-parser": "1.0.0",
+    "@angular/cli": "~11.0.5",
+    "@angular/compiler-cli": "~11.0.5",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "^4.4.1",
-    "@typescript-eslint/parser": "^4.4.1",
-    "codelyzer": "^5.1.2",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
+    "codelyzer": "^6.0.0",
+    "eslint": "^7.6.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
     "husky": "^4.3.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "lint-staged": "^10.4.2",
     "markdownlint-cli": "^0.24.0",
-    "ng-packagr": "^10.1.0",
-    "npm-run-all": "^4.1.5",
+    "ng-packagr": "^11.0.3",
     "protractor": "~7.0.0",
     "standardx": "^5.0.0",
     "stylelint": "^13.7.2",
