@@ -1,14 +1,22 @@
 import { InjectionToken, Provider } from '@angular/core'
-import { switchMap, tap, publishReplay, refCount, catchError } from 'rxjs/operators'
+import { switchMap, publishReplay, refCount, catchError } from 'rxjs/operators'
 import { ActivatedRoute } from '@angular/router'
 import { LoadingWrapper, LoadingWrapperModel } from '@libs/loading-wrapper/loading-wrapper'
-import { ContractGrantModel } from '@services/contract/contract.model'
+import {
+  ContractGrantModel
+} from '@services/contract/contract.model'
 import { ContractService } from '@services/contract/contract.service'
 import { MatSnackBar } from '@angular/material/snack-bar'
 import { translate } from '@ngneat/transloco'
+import { ContractProviderDefine } from '@services/contract/contract-provider-factory'
+import { GrantsVariationType } from '@services/static/static.model'
 
 export const ENTITY = new InjectionToken<LoadingWrapperModel<ContractGrantModel>>(
   'A stream with current contract'
+)
+
+export const CONTRACT = new InjectionToken<GrantsVariationType>(
+  'A stream with contract info'
 )
 
 // По этому токену будет идти стрим с необходимой компоненту информацией:
@@ -17,7 +25,8 @@ export const ENTITY_PAGE_PROVIDERS: Provider[] = [
     provide: ENTITY,
     deps: [ContractService, ActivatedRoute, MatSnackBar],
     useFactory: EntityFactory
-  }
+  },
+  ContractProviderDefine(CONTRACT)
 ]
 
 export function EntityFactory (
