@@ -1,25 +1,47 @@
 import { Observable, Subject, EMPTY } from 'rxjs'
-import { shareReplay, catchError } from 'rxjs/operators'
+import {
+  catchError,
+  takeUntil,
+  publishReplay,
+  refCount
+} from 'rxjs/operators'
 
 export interface LoadingWrapperModel<T> {
   data$: Observable<T>
   error$: Observable<boolean>
+  destroy: () => void
 }
 
 export class LoadingWrapper<T> {
   private readonly errorLoading$ = new Subject<boolean>()
-  readonly error$: Observable<boolean> = this.errorLoading$.pipe(shareReplay(1))
   readonly data$: Observable<T>
+
+  private destroyed$ = new Subject();
+
+  readonly error$: Observable<boolean> = this.errorLoading$.pipe(
+    takeUntil(this.destroyed$),
+    publishReplay(1),
+    refCount()
+  )
 
   constructor (data: Observable<T>) {
     this.data$ = data.pipe(
-      shareReplay(1),
+      takeUntil(this.destroyed$),
+      publishReplay(1),
+      refCount(),
       catchError((error) => {
         // Todo решить где и как обрабатывать ошибку
         console.log('LoadingWrapper::Error\n', error)
-        this.errorLoading$.next(true)
+        setTimeout(() => {
+          this.errorLoading$.next(true)
+        })
         return EMPTY
       })
     )
+  }
+
+  destroy () {
+    this.destroyed$.next(null)
+    this.destroyed$.complete()
   }
 }
