@@ -2,87 +2,134 @@ import { Inject, Injectable } from '@angular/core'
 import { Signer, IUserData } from '@waves/signer/'
 import Provider from '@waves.exchange/provider-web'
 import { API, AppApiInterface } from '@constants'
-import { SignerUser, SignerInvokeArgs } from './signer.model'
-import { BehaviorSubject, from, Observable, ReplaySubject, Subject } from 'rxjs'
-import { publishReplay, refCount, startWith } from 'rxjs/operators'
-import { UserService } from '@services/user/user.service'
-import { IWithApiMixin, IInvokeScriptTransaction } from '@waves/ts-types'
-import { IMoney } from '@waves/signer/cjs/interface'
-import { InvokeResponseInterface } from '../../interface'
+import {
+  SignerUser,
+  SignerInvokeArgs,
+  TransactionRawState,
+  TransactionState,
+  TransactionsSuccessResult
+} from './signer.model'
+import { BehaviorSubject, from, Observable } from 'rxjs'
+import { publishReplay, refCount, tap, switchMap, retryWhen, delay, map, take } from 'rxjs/operators'
+import {
+  TTransactionFromAPI
+} from '@waves/ts-types'
+import {
+  IMoney, TLong
+} from '@waves/signer/cjs/interface'
+import { HttpClient } from '@angular/common/http'
+import { translate } from '@ngneat/transloco'
+import { MatSnackBar } from '@angular/material/snack-bar'
+import { StorageService } from '@services/storage/storage.service'
 
 @Injectable({
   providedIn: 'root'
 })
 export class SignerService {
-  public readonly signer: Signer
+  public readonly signer!: Signer
 
-  private user$: BehaviorSubject<SignerUser> = new BehaviorSubject({
-    name: '',
-    address: '',
-    publicKey: ''
-  })
+  private user$: BehaviorSubject<SignerUser> = new BehaviorSubject({ name: '', address: '', publicKey: '', balance: '' })
 
-  public user: Observable<SignerUser> = this.user$.pipe(
-    publishReplay(1),
-    refCount()
-
-  )
+  public user: Observable<SignerUser> = this.user$.pipe(tap((data) => {
+    this.storageService.userData = data
+  }), publishReplay(1), refCount())
 
   constructor (
-      @Inject(API) private readonly api: AppApiInterface
+    @Inject(API) private readonly api: AppApiInterface,
+    private readonly http: HttpClient,
+    private snackBar: MatSnackBar,
+    private storageService: StorageService
   ) {
     this.signer = new Signer({
       // Specify URL of the node on Testnet
       NODE_URL: api.nodes
     })
     this.signer.setProvider(new Provider(api.signer))
+
+    this.user$.next(this.storageService.userData as SignerUser)
   }
 
   public login (): Observable<Observable<SignerUser>> {
     return from(
       this.signer.login()
         .then((user: IUserData) => {
-          console.log('login(ser: IUserData):', user)
-          this.user$.next({
-            ...user,
-            name: user.address.slice(0, 3) + '…' + user.address.slice(-3)
-          })
-
+          this.signer.getBalance()
+            .then((res) => {
+              this.user$.next({
+                ...user,
+                balance: res[0].amount.toString(),
+                name: user.address.slice(0, 3) + '…' + user.address.slice(-3)
+              })
+            })
+            .catch((err) => {
+              this.snackBar.open(err, translate('messages.error'))
+              throw new Error(translate('messages.getBalanceError'))
+            })
           return this.user
         })
-        // Todo transform errors to messages in snackbar
     )
   }
 
   public logout (): Observable<void> {
+    this.user$.next({ name: '', address: '', publicKey: '', balance: '' })
     return from(this.signer.logout())
   }
 
-  // @ts-ignore
-  public invoke (command: string, args: SignerInvokeArgs[], payment: Array<IMoney> = []):
-    Promise<[IInvokeScriptTransaction<string | number> & IWithApiMixin]> {
-    const tx = this.signer.invoke({
+  public invokeProcess (
+    contractAddress: string,
+    command: string,
+    args: SignerInvokeArgs[],
+    payment: Array<IMoney> = []
+  ): Observable<TransactionsSuccessResult> {
+    return from(this.signer.invoke({
       payment,
-      dApp: this.api.contractAddress,
+      dApp: contractAddress,
       call: {
         function: command,
         // @ts-ignore
         args
       }
-    })
+    }).sign()).pipe(
+      take(1),
+      tap(() => {
+        this.snackBar.open(translate('messages.startTransaction'), translate('messages.ok'))
+      }),
+      // @ts-ignore
+      switchMap((tx) => {
+        return from(this.signer.broadcast(tx))
+      }),
+      switchMap((data: TTransactionFromAPI<TLong>) => {
+        return this.status(data.id)
+      }),
+      tap(() => {
+        this.snackBar.open('Transaction is complete', translate('messages.ok'))
+      })
+    )
+  }
 
-    // let v = this.signer.broadcast(tx.id, {chain: true, confirmations: 1})
-    // signer.invoke({
-    //   dApp: address,
-    //   call: { function: name, args: convertedArgs },
-    // }).sign();
-    // console.log('------>', tx.alias().)
-    // this.signer.waitTxConfirm(tx.alias, {}).then((e)=>{
-    //   console.log('!!!', e)
-    // })
+  status (tx: string): Observable<TransactionsSuccessResult> {
+    const url = new URL('/transactions/status/', this.api.rest)
+    return this.http.get<TransactionRawState>(url.href, {
+      params: {
+        id: tx
+      },
+      headers: { accept: 'application/json; charset=utf-8' }
+    }).pipe(
+      map((data: TransactionState[]) => {
+        const confirmation = data.find((state: TransactionState) => {
+          return state.status === 'confirmed' && state.confirmations >= this.api.confirmations
+        })
 
-    return tx.broadcast(
-      // {chain: true, confirmations: 1}
+        console.log('Confirmation', confirmation)
+        if (!confirmation) {
+          throw new Error('wait')
+        }
+
+        return confirmation as TransactionsSuccessResult
+      }),
+      retryWhen((data) => {
+        return data.pipe(delay(1000))
+      })
     )
   }
 }
