@@ -1,11 +1,12 @@
-import { AppPage } from './app.po'
+
 import { browser, logging } from 'protractor'
 
 describe('workspace-project App', () => {
-  let page: AppPage
+  // let page: AppPage
 
   beforeEach(() => {
-    page = new AppPage()
+    // page = new AppPage()
+    // new AppPage()
   })
 
   // it('should display title in header', () => {
