@@ -15,24 +15,14 @@ const routes: Routes = [
     component: CommonLayoutComponent,
     children: [
       {
-        path: environment.routing.listing,
-        loadChildren: () => import('./pages/listing-page/listing-page.module').then((m) => m.ListingPageModule),
+        path: environment.routing.home,
+        loadChildren: () => import('./pages/home-page/home-page.module').then((m) => m.HomePageModule),
         pathMatch: 'full'
       },
       {
-        path: environment.routing.entity,
-        loadChildren: () =>
-          import('./pages/entity-page/entity-page.module').then((m) => m.EntityPageModule)
-      },
-      {
-        path: environment.routing.application,
-        loadChildren: () =>
-          import('./pages/application-page/application-page.module').then((m) => m.ApplicationPageModule)
-      },
-      {
-        path: environment.routing.masterSetting,
-        loadChildren: () =>
-          import('./pages/master-setting-page/master-setting-page.module').then((m) => m.MasterSettingPageModule)
+        path: environment.routing.listing,
+        loadChildren: () => import('./pages/listing-page/listing-page.module').then((m) => m.ListingPageModule),
+        pathMatch: 'full'
       },
       {
         path: environment.routing.about,
@@ -40,37 +30,38 @@ const routes: Routes = [
         pathMatch: 'full'
       },
       {
-        path: environment.routing.createGrant,
-        loadChildren: () => import('./pages/create-grant-page/create-grant-page.module').then((m) => m.CreateGrantPageModule),
+        path: environment.routing.members,
+        loadChildren: () => import('./pages/members-page/members-page.module').then((m) => m.MembersPageModule),
         pathMatch: 'full'
       },
-      {
-        path: environment.routing.editGrant,
-        loadChildren: () => import('./pages/edit-grant-page/edit-grant-page.module').then((m) => m.EditGrantPageModule),
-        pathMatch: 'full'
-      },
-      {
-        path: environment.routing.applyGrant,
-        loadChildren: () => import('./pages/apply-grant-page/apply-grant-page.module').then((m) => m.ApplyGrantPageModule),
-        pathMatch: 'full'
-      },
-      {
-        path: environment.routing.setting,
-        loadChildren: () => import('./pages/setting-page/setting-page.module').then((m) => m.SettingPageModule),
-        pathMatch: 'full'
-      },
-      {
-        path: environment.routing.create,
-        loadChildren: () => import('./pages/create-entity-page/create-entity-page.module').then((m) => m.CreateEntityPageModule),
-        pathMatch: 'full'
-      },
+      // {
+      //   path: environment.routing.createGrant,
+      //   loadChildren: () => import('./pages/create-grant-page/create-grant-page.module').then((m) => m.CreateGrantPageModule),
+      //   pathMatch: 'full'
+      // },
+      // {
+      //   path: environment.routing.editGrant,
+      //   loadChildren: () => import('./pages/edit-grant-page/edit-grant-page.module').then((m) => m.EditGrantPageModule),
+      //   pathMatch: 'full'
+      // },
       {
         path: environment.routing.stylesheet,
         loadChildren: () => import('./pages/stylesheet-page/stylesheet-page.module').then((m) => m.StylesheetPageModule),
         pathMatch: 'full'
+      },
+      // {
+      //   path: environment.routing.addReward,
+      //   loadChildren: () => import('./pages/add-reward-page/add-reward-page.module').then((m) => m.AddRewardPageModule),
+      //   pathMatch: 'full'
+      // },
+      {
+        path: environment.routing.entity,
+        loadChildren: () =>
+          import('./pages/entity-page/entity-page.module').then((m) => m.EntityPageModule)
       }
     ]
   },
+
   {
     path: '**',
     loadChildren: () =>
