@@ -3,6 +3,8 @@ import { RouterTestingModule } from '@angular/router/testing'
 import { DevGridModule } from '@ui/dev-grid/dev-grid.module'
 import { AppComponent } from './app.component'
 import { NgProgressModule } from 'ngx-progressbar'
+import { MarkdownModule } from 'ngx-markdown'
+import { NoopAnimationsModule } from '@angular/platform-browser/animations'
 
 describe('AppComponent', () => {
   beforeEach(async () => {
@@ -10,7 +12,9 @@ describe('AppComponent', () => {
       imports: [
         RouterTestingModule,
         DevGridModule,
-        NgProgressModule
+        NgProgressModule,
+        MarkdownModule,
+        NoopAnimationsModule
       ],
       declarations: [
         AppComponent
