@@ -2,14 +2,43 @@ import { InjectionToken, Provider, Type } from '@angular/core'
 import { environment } from '../environments/environment'
 
 export interface AppConstantsInterface {
+  production: boolean
   routes: {[s: string]: string | string[]}
+}
+
+export interface UserInterface {
+  name: string,
+  twitter?: string,
+  facebook?: string,
+  linkedin?: string
 }
 
 export interface AppApiInterface {
   nodes: string
   signer: string
   rest: string
-  contractAddress: string
+  explorer: string
+  confirmations: number
+  management: {
+    membership: string,
+  },
+  contracts: {
+    disruptive: string,
+    web3: string,
+    interhack: string
+  },
+  grantsProgramLink: string,
+  issues: {
+    disruptive: string,
+    web3: string,
+    interhack: string
+  },
+  workingGroup: {[s: string]: {
+    name: string,
+    twitter?: string,
+    facebook?: string,
+    linkedin?: string
+  }}
 }
 
 export const APP_CONSTANTS = new InjectionToken<Type<AppConstantsInterface>>('Application constants')
@@ -21,7 +50,10 @@ export function provideApi (): Provider[] {
     {
       provide: API,
       useValue: {
-        ...environment.apis
+        ...environment.apis,
+        confirmations: environment.confirmations,
+        grantsProgramLink: environment.grantsProgramLink,
+        workingGroup: environment.workingGroup
       }
     }
   ]
@@ -32,7 +64,8 @@ export function provideAppConstants (): Provider[] {
     {
       provide: APP_CONSTANTS,
       useValue: {
-        routes: environment.routing
+        routes: environment.routing,
+        production: environment.production
       }
     }
   ]
