@@ -1,6 +1,15 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { AboutPageComponent } from './about-page.component'
+import { getTranslocoModule } from '@dapp/src/app/transloco-module.spec'
+import { HttpClientTestingModule } from '@angular/common/http/testing'
+import { RouterTestingModule } from '@angular/router/testing'
+import { provideApi, provideAppConstants } from '@constants'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
+import { ContractProviderDefine } from '@services/contract/contract-provider-factory'
+import { CONTRACT } from './about-page.provider'
+import { NotFoundPageModule } from '@pages/not-found-page/not-found-page.module'
+import { LoadingPageModule } from '@pages/loading-page/loading-page.module'
 
 describe('AboutPageComponent', () => {
   let component: AboutPageComponent
@@ -8,7 +17,20 @@ describe('AboutPageComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [AboutPageComponent]
+      imports: [
+        HttpClientTestingModule,
+        getTranslocoModule(),
+        RouterTestingModule,
+        MatSnackBarModule,
+        NotFoundPageModule,
+        LoadingPageModule
+      ],
+      declarations: [AboutPageComponent],
+      providers: [
+        provideAppConstants(),
+        provideApi(),
+        ContractProviderDefine(CONTRACT)
+      ]
     })
       .compileComponents()
   })
