@@ -1,10 +1,16 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
-
 import { ListingPageComponent } from './listing-page.component'
 import { ContractService } from '@services/contract/contract.service'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
-import { provideApi } from '@constants'
-import { UserService } from '@services/user/user.service'
+import { provideApi, provideAppConstants } from '@constants'
+import { ListingModule } from '@ui/listing/listing.module'
+import { RouterTestingModule } from '@angular/router/testing'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
+import { getTranslocoModule } from '@dapp/src/app/transloco-module.spec'
+import { LoadingPageModule } from '@pages/loading-page/loading-page.module'
+import { NoopAnimationsModule } from '@angular/platform-browser/animations'
+import { ContractProviderDefine } from '@services/contract/contract-provider-factory'
+import { CONTRACT } from './listing-page.providers'
 
 describe('ListingPageComponent', () => {
   let component: ListingPageComponent
@@ -12,9 +18,22 @@ describe('ListingPageComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
+      imports: [
+        HttpClientTestingModule,
+        RouterTestingModule,
+        ListingModule,
+        MatSnackBarModule,
+        getTranslocoModule(),
+        LoadingPageModule,
+        NoopAnimationsModule
+      ],
       declarations: [ListingPageComponent],
-      providers: [ContractService, provideApi()]
+      providers: [
+        provideAppConstants(),
+        ContractService,
+        provideApi(),
+        ContractProviderDefine(CONTRACT)
+      ]
     })
       .compileComponents()
   })
