@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { NotFoundPageComponent } from './not-found-page.component'
-import { TranslocoModule } from '@ngneat/transloco'
+import { getTranslocoModule } from '../../transloco-module.spec'
 
 describe('NotFoundPageComponent', () => {
   let component: NotFoundPageComponent
@@ -9,7 +9,7 @@ describe('NotFoundPageComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [TranslocoModule],
+      imports: [getTranslocoModule()],
       declarations: [NotFoundPageComponent]
     })
       .compileComponents()
