@@ -4,10 +4,12 @@ import { CommonModule } from '@angular/common'
 import { NotFoundPageRoutingModule } from './not-found-page-routing.module'
 import { NotFoundPageComponent } from './not-found-page.component'
 import { TranslocoModule } from '@ngneat/transloco'
-import { CreateGrantPageGuard } from '@pages/create-grant-page/create-grant-page.guard'
 
 @NgModule({
   declarations: [NotFoundPageComponent],
+  exports: [
+    NotFoundPageComponent
+  ],
   imports: [
     CommonModule,
     NotFoundPageRoutingModule,
