@@ -23,4 +23,4 @@ getTestBed().initTestEnvironment(
 // Then we find all the tests.
 const context = require.context('./', true, /\.spec\.ts$/)
 // And load the modules.
-context.keys().map(context)
+context.keys().forEach(context)
