@@ -1,16 +1,19 @@
 import { TestBed } from '@angular/core/testing'
 import { SignerService } from './signer.service'
-import { provideApi } from '@constants'
+import { provideApi, provideAppConstants } from '@constants'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
+import { RouterTestingModule } from '@angular/router/testing'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
 
 describe('SignerService', () => {
   let service: SignerService
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
+      imports: [HttpClientTestingModule, RouterTestingModule, MatSnackBarModule],
       providers: [
-        provideApi()
+        provideApi(),
+        provideAppConstants()
       ]
     })
     service = TestBed.inject(SignerService)
