@@ -1,6 +1,11 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { FooterComponent } from './footer.component'
+import { HttpClientTestingModule } from '@angular/common/http/testing'
+import { provideApi, provideAppConstants } from '@constants'
+import { RouterTestingModule } from '@angular/router/testing'
+import { getTranslocoModule } from '@dapp/src/app/transloco-module.spec'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
 
 describe('FooterComponent', () => {
   let component: FooterComponent
@@ -8,7 +13,12 @@ describe('FooterComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [FooterComponent]
+      imports: [getTranslocoModule(), HttpClientTestingModule, RouterTestingModule, MatSnackBarModule],
+      declarations: [FooterComponent],
+      providers: [
+        provideAppConstants(),
+        provideApi()
+      ]
     })
       .compileComponents()
   })
