@@ -4,7 +4,6 @@ import { HeaderComponent } from './header.component'
 import { TranslocoModule } from '@ngneat/transloco'
 import { PipesModule } from '@libs/pipes/pipes.module'
 import { RouterModule } from '@angular/router'
-import { PopupModule } from '@ui/popup/popup.module'
 
 @NgModule({
   declarations: [HeaderComponent],
@@ -12,8 +11,7 @@ import { PopupModule } from '@ui/popup/popup.module'
     CommonModule,
     TranslocoModule,
     PipesModule,
-    RouterModule,
-    PopupModule
+    RouterModule
   ],
   exports: [HeaderComponent]
 })
