@@ -5,6 +5,8 @@ import { RouterModule } from '@angular/router'
 import { PipesModule } from '@libs/pipes/pipes.module'
 import { TranslocoModule } from '@ngneat/transloco'
 import { TagModule } from '@ui/tag/tag.module'
+import { SubListModule } from '@ui/listing/sub-list/sub-list.module'
+import { MatDialogModule } from '@angular/material/dialog'
 
 @NgModule({
   declarations: [ListingComponent],
@@ -13,7 +15,9 @@ import { TagModule } from '@ui/tag/tag.module'
     RouterModule,
     TranslocoModule,
     PipesModule,
-    TagModule
+    TagModule,
+    SubListModule,
+    MatDialogModule
   ],
   exports: [ListingComponent]
 })
