@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Inject, Input, OnInit, Output } from '@angular/core'
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core'
 
 @Component({
   selector: 'ui-tag',
@@ -6,12 +6,19 @@ import { ChangeDetectionStrategy, Component, EventEmitter, Inject, Input, OnInit
   styleUrls: ['./tag.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TagComponent {
+export class TagComponent implements OnInit {
   @Input() tags: string[] = []
   @Output() selectedTag = new EventEmitter<string>();
   lastTag = ''
+
+  constructor () {
+  }
+
+  ngOnInit (): void {
+    this.select('all')
+  }
+
   select (tag: string): void {
-    console.log(tag, this.lastTag, tag === this.lastTag)
     if (this.lastTag === tag) {
       this.lastTag = ''
       this.selectedTag.emit('')
